Change to the logic around Disable Server Verification
Currently, using the `-dsv True` flag is over-ridden by the `-certs` flag, which, if not provided manually is provided by default.

So, before this change, specifying `-dsv True` without specifying an empty set of `-certs` will *not* disable server verification. 
This is counter-intuitive - if I specify dsv, then I expect server verification to be off completely.
As a side note - while it is easy enough on linux to specify an empty set of certs ( `-certs ""`), I have seen that this does not hold true for all OS - e.g. on Windows it is not straightforward, it seems.

Steps to reproduce:
- run the script against a TLS enabled Dremio whose cert is not trusted by the system the script is running on, and include `-dsv True` but do not include a certs parameter.
Result:
- Script will fail with TLS errors because the system certs are being used which do not trust the Dremio cert